visual studio 2008 build project for cuda kernels
Ignore-this: 559f90bbc026b01b0258621fa859da3
- Just as the makefiles will, this places kernels in dist/lib

darcs-hash:20100406052409-9241b-1dc7bee89b0430013ec16dbc3d107f7df1926c80.gz

+++ src/common/include/utils.h
@@ -115,7 +115,7 @@ floorPow2(unsigned int x)
 #endif
 
     int exp;
-    frexp(x, &exp);
+    frexp((double)x, &exp);
     return 1 << (exp - 1);
 }
